Deleted reference to nonexistent example
The old documentation contained a reference to Example D, "The Facts Were These, Refactored." This is not included in the documentation but only exists in a post on intfiction.org, linked earlier in the documentation. I deleted the reference to it as Example D and added a tiny bit more of a description where the link goes.